+++ ChangeLog.md
@@ -1,4 +1,5 @@
 ### 5.2.5 (TBD)
+- Add support for constructing ImageData/VolumeData from multi-frame datasets (#2015)
 - Added some private tags to dictionary (#2027)
 
 ### 5.2.4 (2025-10-03)